% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axi_cmd.R
\name{axi_exec}
\alias{axi_exec}
\alias{axi_version}
\alias{axi_sysinfo}
\alias{axi_toggle_pen}
\alias{axi_align}
\alias{axi_pen_test}
\title{Execute direct commands to the AxiDraw plotter}
\usage{
axi_version(options = axi_options())

axi_sysinfo(options = axi_options())

axi_toggle_pen(options = axi_options())

axi_align(options = axi_options())

axi_pen_test(tip_size, line_overlap, options = axi_options())
}
\arguments{
\item{options}{An \code{axi_options} object. See the documentation for
\code{\link[=axi_options]{axi_options()}} for all the settings. Many are not relevant for these
function.}

\item{tip_size}{The size of the tip (i.e. the width of the line it draws) of
the initial pen}

\item{line_overlap}{The overlap between adjacent pen strokes when filling out
shapes and drawing thick lines, in mm. Setting this to a negative amount will
cause gaps between the lines.}
}
\description{
These small utility functions are not concerned with plotting, but rather
lets you query information about the plotter or execute small commands useful
for setup.
}
\details{
The different functions have various, unrelated, uses:
\itemize{
\item \code{axi_version()} returns the version of the AxiDraw software
\item \code{axi_sysinfo()} returns the EBB firmware version and system information
\item \code{axi_toggle_pen()} toggles the position of the pen (raises it if lowered
and vice versa)
\item \code{axi_align()} raises the pen and turns off the stepper motors so the
carriage can be moved freely.
}
}
