% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{axi_svg}
\alias{axi_svg}
\title{Render an SVG file with the AxiDraw penplotter}
\usage{
axi_svg(file, options = list(), capture = FALSE, text)
}
\arguments{
\item{file}{An svg file}

\item{options}{A list of options for the AxiDraw. Use \code{\link[=svg_options]{svg_options()}} for
help with creating a valid list of options.}

\item{capture}{Logical. Should the drawing instructions be captured and
returned as a new svg.}

\item{text}{An svg as a character vector.}
}
\value{
If \code{capture = TRUE} a new svg showing the pen plotter movement that
will be used for drawing.
}
\description{
This function takes an SVG representation of a graphic and sends it to the
AxiDraw penplotter for drawing. This function uses the build-in svg drawing
capabilities of the pyaxidraw library so rendering an svg with
\code{\link[grDevices:svg]{grDevices::svg()}} and then using \code{axi_svg} to draw it will likely produce
a different result than rendering directly to \code{\link[=axi_dev]{axi_dev()}}.
}
