% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fawkes-package.R
\docType{package}
\name{fawkes-package}
\alias{fawkes}
\alias{fawkes-package}
\title{fawkes: An R Interface to the AxiDraw Plotter}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides interface to the AxiDraw pen plotter through the python API. 'fawkes' contain both a graphic device that allows you to control the plotter through the R graphic engine, as well as facilities to control it directly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fawkes.data-imaginist.com}
  \item \url{https://github.com/thomasp85/fawkes}
  \item Report bugs at \url{https://github.com/thomasp85/fawkes/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
