% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pens.R
\name{pen}
\alias{pen}
\title{Describe the characteristics of a pen}
\usage{
pen(color, tip_size = NULL, offset = c(0, 0), options = NULL)
}
\arguments{
\item{color}{The color that this pen corresponds to.}

\item{tip_size}{The size of the line the pen leaves on the paper in mm}

\item{offset}{The x and y offset from the starting pen tip}

\item{options}{An \code{axi_options} object. See the documentation for
\code{\link[=axi_options]{axi_options()}} for all the settings.}
}
\value{
An \code{axi_pen} object
}
\description{
Pen plots can include ink from many pens and how they behave may differ.
While some of these can be set when a new color is encountered it is often
better and less error prone to define them upfront. It further allows you
to specify specific axidraw options for each pen, e.g. travel speed.
}
\examples{
pen(color = 'forestgreen', tip_size = 2)

# Combine multiple pens
c(
  pen(color = 'forestgreen', tip_size = 2),
  pen(color = 'steelblue', tip_size = 1,
      options = axi_options(speed_down = 15))
)

}
