% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{axi_options}
\alias{axi_options}
\alias{svg_options}
\alias{manual_options}
\title{Setting options for AxiDraw}
\usage{
svg_options(...)

manual_options(...)
}
\arguments{
\item{...}{A range of options}
}
\description{
These functions help with setting options for the AxiDraw. They output
regular lists, but will help with input checks and the likes, and allow
tab-completion of option names.
}
