% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{axi_svg}
\alias{axi_svg}
\title{Render an SVG file with the AxiDraw penplotter}
\usage{
axi_svg(
  file,
  capture = FALSE,
  options = axi_options(),
  layer = NULL,
  copies = 1L,
  page_delay = 15,
  auto_rotate = TRUE,
  rendering = "all",
  reorder = "none",
  summary = FALSE,
  text
)
}
\arguments{
\item{file}{An svg file}

\item{capture}{Logical. Should the drawing instructions be captured and
returned as a new svg.}

\item{options}{An \code{axi_options} object. See the documentation for
\code{\link[=axi_options]{axi_options()}} for all the settings.}

\item{layer}{The layer to plot when \code{mode = 'layers'}. See the details
section}

\item{copies}{The number of copies to produce}

\item{page_delay}{The delay in seconds between each copy}

\item{auto_rotate}{Automatically rotate the image to maximize paper usage}

\item{rendering}{The type of preview to generate when \code{preview = TRUE}.
Either \code{'none'} (nothing rendered), \code{'down'} (render when pen is down),
\code{'up'} (render when pen is up), or \code{'all'} (render everything)}

\item{reorder}{The type of automatical line reordering to perform. Either
\code{'none'} (no reordering), \code{'objects'} (reorder within objects), \code{'groups'} (
reorder within objects, then reorder groups), or \code{'full'} (reorder everything
together).}

\item{summary}{Should time spend and distance travel be reported in the end}

\item{text}{An svg as a character vector.}
}
\value{
If \code{capture = TRUE} a new svg showing the pen plotter movement that
will be used for drawing.
}
\description{
This function takes an SVG representation of a graphic and sends it to the
AxiDraw penplotter for drawing. This function uses the build-in svg drawing
capabilities of the pyaxidraw library so rendering an svg with
\code{\link[grDevices:svg]{grDevices::svg()}} and then using \code{axi_svg} to draw it will likely produce
a different result than rendering directly to \code{\link[=axi_dev]{axi_dev()}}.
}
\examples{
\dontrun{
logo <- system.file('example.svg', package = 'fawkes')
axi_svg(logo)
}

}
