% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{install_axidraw}
\alias{install_axidraw}
\title{Install the AxiDraw python API}
\usage{
install_axidraw(path = NULL, ...)
}
\arguments{
\item{path}{The path to the AxiDraw API. If NULL it will be downloaded
automatically.}

\item{...}{Arguments passed on to \code{reticulate::py_install}
\describe{
  \item{envname}{Name of environment to install packages into}
  \item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}
  \item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}
}}
}
\description{
This function will download and install the AxiDraw python API for use with
reticulate. It is necessary to have this installed in order to use any of the
functionality in the fawkes package.
}
