% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{install_axidraw}
\alias{install_axidraw}
\title{Install the AxiDraw python API}
\usage{
install_axidraw(path = NULL, ...)
}
\arguments{
\item{path}{The path to the AxiDraw API. If NULL it will be downloaded
automatically.}

\item{...}{
  Arguments passed on to \code{\link[reticulate:py_install]{reticulate::py_install}}
  \describe{
    \item{\code{envname}}{The name, or full path, of the environment in which Python
packages are to be installed. When \code{NULL} (the default), the active
environment as set by the \code{RETICULATE_PYTHON_ENV} variable will be used;
if that is unset, then the \code{r-reticulate} environment will be used.}
    \item{\code{method}}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}
    \item{\code{conda}}{The path to a \code{conda} executable. Use \code{"auto"} to allow \code{reticulate} to
automatically find an appropriate \code{conda} binary. See \strong{Finding Conda} for more details.}
    \item{\code{python_version}}{The requested Python version. Ignored when attempting
to install with a Python virtual environment.}
    \item{\code{pip}}{Boolean; use \code{pip} for package installation? This is only relevant
when Conda environments are used, as otherwise packages will be installed
from the Conda repositories.}
  }}
}
\description{
This function will download and install the AxiDraw python API for use with
reticulate. It is necessary to have this installed in order to use any of the
functionality in the fawkes package.
}
